# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Oru'el GPU Relay API
  version: "1.0.0"
  description: |
    Oru'el's relay to the Spheron AI GPU network. Every endpoint here mirrors the
    shape of Spheron's own API (https://docs.spheron.ai/api-reference) field-for-field,
    so an integration built against Spheron's docs needs only two changes to work here:

    1. Point the base URL at Oru'el's relay instead of `https://app.spheron.ai`.
    2. Authenticate with `X-API-Key: <key Oru'el issued you>` instead of a Spheron
       bearer token.

    **Pricing note:** every dollar figure returned by this API (offer prices, hourly
    rates, total cost, volume storage rates, etc.) already includes Oru'el's markup.
    There is nothing further to add on your side - the number you see is the number
    you're billed.

    **Scoping note:** `GET /deployments` and `GET /deployments/{id}` only ever return
    deployments launched with *your* API key. You cannot see or affect another
    partner's instances.

    Endpoints intentionally not exposed here (and why): `GET /balance` and
    `GET /teams` from Spheron's own API expose Oru'el's internal account balance and
    team membership - that's not relayed to any downstream partner.
  contact:
    name: Oru'el
...
servers:
  - url: https://relay.oruel.example/api
    description: Production (replace with your real deployed hostname)
  - url: http://localhost:4000/api
    description: Local development
security:
  - ApiKeyAuth: []
...
paths:
...
  /deployments:
    post:
      tags: [Deployments]
      summary: Launch a GPU instance
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateDeploymentRequest" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Deployment" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "429": { $ref: "#/components/responses/RateLimited" }
      x-rate-limit: "10 requests / 15 min, scoped to your API key"
    get:
      tags: [Deployments]
      summary: List your deployments
      description: Only returns deployments launched with your own API key.
      parameters:
        - name: status
          in: query
          schema: { type: string, enum: [active, inactive, running, deploying, terminated, failed, terminated-provider] }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Deployment" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /deployments/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string } }
    get:
      tags: [Deployments]
      summary: Get deployment details
      description: Refreshes live status/cost from the upstream provider before responding.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Deployment" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      tags: [Deployments]
      summary: Rename a deployment
      description: Only `name` is mutable. Every other attribute is immutable after creation - terminate and relaunch to change anything else.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties: { name: { type: string } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Deployment" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [Deployments]
      summary: Terminate a deployment
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string, example: "Instance destruction initiated" }
                  deployment:
                    type: object
                    properties:
                      id: { type: string }
                      status: { type: string, example: terminated }
                      stoppedAt: { type: string, format: date-time }
        "404": { $ref: "#/components/responses/NotFound" }
  /deployments/{id}/can-terminate:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string } }
    get:
      tags: [Deployments]
      summary: Check whether a deployment can be terminated yet
      description: Call before DELETE to avoid a rejected request if the provider enforces a minimum runtime.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  canTerminate: { type: boolean }
                  runtime: { type: integer, description: "Minutes since the deployment started" }
                  timeRemaining: { type: integer }
                  minimumRuntime: { type: integer }
                  reason: { type: string, description: "Present only when canTerminate is false" }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /ssh-keys:
    get:
      tags: [SSH Keys]
      summary: List your SSH keys
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/SshKey" }
    post:
      tags: [SSH Keys]
      summary: Add an SSH key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, publicKey]
              properties:
                name: { type: string }
                publicKey: { type: string }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SshKey" }
  /ssh-keys/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string } }
    get:
      tags: [SSH Keys]
      summary: Get SSH key details
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SshKey" }
    delete:
      tags: [SSH Keys]
      summary: Delete an SSH key
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { type: object, properties: { success: { type: boolean } } }
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: The key Oru'el issued you. Never share it or embed it in client-side code.
  responses:
    Unauthorized:
      description: Missing or invalid API key
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: Resource does not exist, or does not belong to you
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    BadRequest:
      description: Invalid parameters or request body
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    RateLimited:
      description: Too many requests
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    Error:
      type: object
      properties:
        error: { type: string }
        code: { type: string }
        details: { type: object }
...
    CreateDeploymentRequest:
      type: object
      required: [provider, offerId, gpuType, gpuCount, region, operatingSystem, instanceType]
      properties:
        provider: { type: string, example: spheron-ai }
        offerId: { type: string }
        gpuType: { type: string, example: rtx-4090 }
        gpuCount: { type: integer, example: 1 }
        region: { type: string, example: us-east-1 }
        operatingSystem: { type: string, example: ubuntu-22.04 }
        instanceType: { type: string, enum: [SPOT, DEDICATED, CLUSTER] }
        sshKeyId: { type: string, description: "Either this or ssh_public_key is required" }
        ssh_public_key: { type: string, description: "Inline key content; a temporary key is created if used instead of sshKeyId" }
        name: { type: string }
        cloudInit:
          type: object
          properties:
            runcmd: { type: array, items: { type: string } }
            packages: { type: array, items: { type: string } }
            writeFiles:
              type: array
              items:
                type: object
                properties:
                  path: { type: string }
                  content: { type: string }
                  owner: { type: string }
                  permissions: { type: string }
        kubernetesAddon:
          type: object
          description: "Voltage Park bare-metal CLUSTER instances only"
          properties:
            version: { type: string, example: "1.35" }
            authentication_config_b64: { type: string }
        volumeIds:
          type: array
          items: { type: string }
          description: "Per-provider attach limits apply - see /volumes/regions"
    Deployment:
      type: object
      properties:
        id: { type: string }
        name: { type: string, nullable: true }
        providerId: { type: string }
        gpuType: { type: string }
        offerId: { type: string }
        gpuCount: { type: integer }
        region: { type: string }
        operatingSystem: { type: string }
        instanceType: { type: string }
        ipAddress: { type: string, nullable: true }
        status: { type: string, enum: [deploying, running, failed, terminated, terminated-provider] }
        startedAt: { type: string, format: date-time, nullable: true }
        stoppedAt: { type: string, format: date-time, nullable: true }
        totalCost: { type: number, format: float, description: "USD accrued so far, markup already applied" }
        hourlyRate: { type: number, format: float, description: "USD/hr, markup already applied" }
        vcpus: { type: integer }
        memory: { type: integer }
        storage: { type: integer }
        sshCommand: { type: string, nullable: true }
        sshPort: { type: integer, nullable: true }
        createdAt: { type: string, format: date-time }
    SshKey:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        publicKey: { type: string }
        fingerprint: { type: string }
        createdAt: { type: string, format: date-time }
...